+++ app/src/main/java/com/app/arcabyolimpo/data/remote/api/ArcaApi.kt
@@ -9,6 +9,7 @@ import com.app.arcabyolimpo.data.remote.dto.beneficiaries.BeneficiariesListDto
 import com.app.arcabyolimpo.data.remote.dto.beneficiaries.BeneficiaryDto
 import com.app.arcabyolimpo.data.remote.dto.beneficiaries.GetBeneficiariesDisabilitiesDto
 import com.app.arcabyolimpo.data.remote.dto.disabilities.DisabilityDto
+import com.app.arcabyolimpo.data.remote.dto.disabilities.DisabilityRegisterDto
 import com.app.arcabyolimpo.data.remote.dto.filter.FilterDto
 import com.app.arcabyolimpo.data.remote.dto.password.RecoverPasswordDto
 import com.app.arcabyolimpo.data.remote.dto.password.RecoverPasswordResponseDto
@@ -336,6 +337,11 @@ interface ArcaApi {
     suspend fun getDisabilityDetail(
         @Path("idDisability") id: String,
     ): DisabilityDto
+
+    @POST("discapacity/add")
+    suspend fun registerDisability(
+        @Body requestBody: DisabilityRegisterDto,
+    )
 
     @GET("supplies/workshop/category")
     suspend fun getWorkshopCategoryList(): WorkshopCategoryListDto