+++ app/src/main/java/com/app/arcabyolimpo/data/remote/dto/disabilities/DisabilityRegisterDto.kt
@@ -1,0 +1,6 @@
+package com.app.arcabyolimpo.data.remote.dto.disabilities
+
+data class DisabilityRegisterDto(
+    val nombre: String,
+    val descripcion: String,
+)